Feat-57 카테고리별 게시물 목록 조회 및 게시물 상세 조회

+++ src/main/java/com/gaethering/gaetheringserver/domain/board/dto/PostUpdateResponse.java
@@ -19,7 +19,7 @@ public class PostUpdateResponse {
 	private String title;
 	private String content;
 	private String categoryName;
-	private int viewCnt;
+	private long viewCnt;
 	private int heartCnt;
 	private String nickname;
 	private List<PostImageUrlResponse> imageUrls;
