+++ src/main/java/com/gaethering/gaetheringserver/domain/board/dto/PostWriteResponse.java
@@ -22,7 +22,7 @@ public class PostWriteResponse {
 
     private String categoryName;
 
-    private int viewCnt;
+    private long viewCnt;
 
     private int heartCnt;
